task/change logout logic

+++ src/components/Header.jsx
@@ -1,15 +1,26 @@
 // import React from 'react'
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBell, faHomeUser } from "@fortawesome/free-solid-svg-icons";
+import { faBell, faSignOutAlt } from "@fortawesome/free-solid-svg-icons";
 import PropTypes from "prop-types";
 import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const AppHeader = () => {
   const navigate = useNavigate();
   const handleIconClick = () => {
-    localStorage.clear();
-    navigate("/");
-    window.location.reload();
+    Swal.fire({
+      title: "Are you sure you want to log out?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, log out",
+      cancelButtonText: "Cancel",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        navigate("/");
+        window.location.reload();
+      }
+    });
   };
   return (
     <div className="w-full">
@@ -19,7 +30,7 @@ const AppHeader = () => {
         <FontAwesomeIcon
           className="text-xl px-4"
           onClick={handleIconClick}
-          icon={faHomeUser}
+          icon={faSignOutAlt}
         />
       </div>
     </div>
